Internalize SignDoc and omitDefaults for sign bytes creation
