+++ packages/proto-signing/types/index.d.ts
@@ -1,4 +1,3 @@
-export { omitDefaults } from "./adr27";
 export { Coin } from "./msgs";
 export { cosmosField } from "./decorator";
 export { Registry } from "./registry";
