+++ packages/proto-signing/types/signing.d.ts
@@ -1,2 +1,7 @@
-import { cosmos } from "./generated/codecimpl";
-export declare function makeSignBytes(signDoc: cosmos.tx.ISignDoc): Uint8Array;
+export declare function makeSignBytes(
+  txBody: Uint8Array,
+  authInfo: Uint8Array,
+  chainId: string,
+  accountNumber: number,
+  sequence: number,
+): Uint8Array;
